+++ application/yabam/yabam-core/src/main/java/com/application/presentation/call/controller/CallController.java
@@ -10,11 +10,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.annotation.AssignUserPassport;
+import com.annotation.HasRole;
 import com.application.presentation.call.api.CallApi;
 import com.application.presentation.call.dto.request.CallCreateRequest;
 import com.application.presentation.call.dto.response.CallCursorResponse;
-import com.authorization.AssignUserPassport;
-import com.authorization.HasRole;
 import com.response.ResponseBody;
 import com.vo.UserPassport;
 import com.vo.UserRole;
