+++ application/yabam/yabam-core/src/main/java/com/application/presentation/menu/controller/MenuCategoryController.java
@@ -17,12 +17,12 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.annotation.AssignUserPassport;
+import com.annotation.HasRole;
 import com.application.presentation.menu.api.MenuCategoryApi;
 import com.application.presentation.menu.dto.request.PatchMenuCategoryInfoRequest;
 import com.application.presentation.menu.dto.request.PostMenuCategoryInfoRequest;
 import com.application.presentation.menu.dto.response.MenuCategoryInfoResponse;
-import com.authorization.AssignUserPassport;
-import com.authorization.HasRole;
 import com.response.ResponseBody;
 import com.vo.UserPassport;
