+++ common/mvc/src/main/java/com/annotation/AssignUserPassport.java
@@ -1,4 +1,4 @@
-package com.authorization;
+package com.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
